resort/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.16.1
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.1.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
boto3==1.39.7
botocore==1.39.7
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==0.4.17
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
connectors==4.1
contourpy==1.0.6
cryptography==45.0.5
cv==1.0.0
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.9
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
django-datatable==0.3.1
django-import-export==3.3.8
django-widget-tweaks==1.5.1
djangorestframework==3.15.2
dnspython==2.7.0
dotenv==0.9.9
durationpy==0.10
ecdsa==0.19.1
email_validator==2.2.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
Faker==19.13.0
fastapi==0.104.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fastjsonschema==2.16.2
filelock==3.17.0
firebase-admin==6.2.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fonttools==4.38.0
frozenlist==1.4.1
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.2
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
kubernetes==34.1.0
langchain==0.1.3
langchain-community==0.0.16
langchain-core==0.1.23
langchain-openai==0.0.5
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.0.87
Mako==1.3.10
Markdown==3.5.1
markdown-it-py==4.0.0
MarkupPy==1.14
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mpmath==1.3.0
msal==1.32.3
msgpack==1.1.0
multidict==6.0.5
mypy_extensions==1.1.0
mysql==0.0.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.5
nltk==3.9.2
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.10.2
numpy==1.26.2
oauthlib==3.3.1
odfpy==1.4.1
onnxruntime==1.23.2
openai==0.28.0
opencv-python==4.11.0.86
openpyxl==3.1.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
patsy==0.5.3
pickleshare==0.7.5
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.6.0
posthog==7.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.4
pulsar-client==3.8.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.1.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyodbc==5.2.0
pyparsing==3.0.9
PyPika==0.48.9
pyreadline3==3.5.4
pyrsistent==0.19.2
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.20
pytz==2022.6
pywin32==305
pywinpty==2.0.9
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
razorpay==1.4.2
redis==6.2.0
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.6.2
scikit-learn==1.1.3
scipy==1.16.3
seaborn==0.12.1
selenium==4.33.0
Send2Trash==1.8.0
sendgrid==6.12.4
sentence-transformers==2.2.2
sentencepiece==0.2.1
sentry-sdk==2.41.0
shellingham==1.5.4
six==1.17.0
sklearn==0.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.1
starlette==0.27.0
statsmodels==0.13.5
sympy==1.14.0
tablib==3.5.0
temporalio==1.12.0
tenacity==8.5.0
terminado==0.17.0
threadpoolctl==3.1.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tornado==6.2
tqdm==4.67.1
traitlets==5.5.0
transformers==4.57.1
trio==0.30.0
trio-websocket==0.12.2
twilio==9.2.1
typer==0.19.2
types-protobuf==6.30.2.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.24.0
virtualenv==20.29.2
watchfiles==1.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.11.0
widgetsnbextension==4.0.3
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.23.0
zstandard==0.23.0
